Crate glean_ffi

Functions

glean_boolean_set

glean_boolean_should_record

glean_boolean_test_get_value

glean_boolean_test_has_value

glean_counter_add

glean_counter_should_record

glean_counter_test_get_value

glean_counter_test_has_value

glean_destroy_boolean_metric

glean_destroy_counter_metric

glean_destroy_glean

glean_destroy_labeled_boolean_metric

glean_destroy_labeled_counter_metric

glean_destroy_labeled_string_metric

glean_destroy_ping_type

glean_destroy_string_metric

glean_enable_logging
Initialize the logging system based on the target platform. This ensures that logging is shown when executing glean unit tests.

glean_initialize

glean_is_initialized

glean_is_upload_enabled

glean_labeled_boolean_metric_get
Create a new instance of the sub-metric of this labeled metric.

glean_labeled_counter_metric_get
Create a new instance of the sub-metric of this labeled metric.

glean_labeled_string_metric_get
Create a new instance of the sub-metric of this labeled metric.

glean_new_boolean_metric

glean_new_counter_metric

glean_new_labeled_boolean_metric
Create a new labeled metric.

glean_new_labeled_counter_metric
Create a new labeled metric.

glean_new_labeled_string_metric
Create a new labeled metric.

glean_new_ping_type

glean_new_string_metric

glean_ping_collect

glean_register_ping_type

glean_send_ping

glean_send_ping_by_name

glean_set_upload_enabled

glean_str_free⚠
Public destructor for strings managed by the other side of the FFI.

glean_string_set

glean_string_should_record

glean_string_test_get_value

glean_string_test_has_value

glean_test_has_ping_type