Module glean_ffi::byte_buffer

ByteBuffer is a struct that represents an array of bytes to be sent over the FFI boundaries.

This is a copy of the same struct from ffi-support, with the difference that the length is encoded as a i32.

Structs

ByteBuffer
ByteBuffer is a struct that represents an array of bytes to be sent over the FFI boundaries. There are several cases when you might want to use this, but the primary one for us is for returning protobuf-encoded data to Swift and Java. The type is currently rather limited (implementing almost no functionality), however in the future it may be more expanded.